sns 파일 네이밍 변경 및 sns 메인페이지 변경,  sns 헤드 추가
